Introduced Gliding Power
Refix gliding power to be in the movement script so it can be easily accessed. There is something going on when trying to access the Stamina script that it doesn't want to properly use the values and assign them. Also try to get the stamina bar to be reduced and filled back up.