Level Selecting works (partially), also introduced the scenes for the next couple levels.
Going to work on Levels 1-5 tomorrow and make a Sprite Palette to use in all the different maps so the assets will be solely under my production. Using a free Sprite Map currently, plan on using my own so I can edit and know more about the Tilemap system in Unity.